+++ src/error/mod.rs
@@ -63,13 +63,6 @@ pub enum Error {
 
     ModuleLoadErrors(collections::HashMap<ModulePath, Error>),
 
-    /// An error encountered while serializing the compiled module to an object file.
-    #[deprecated]
-    ObjectCodeError {
-        filename: path::PathBuf,
-        message: String,
-    },
-
     /// An error ecountered reaading or doing an initial parse of the source code.
     SourceCodeError {
         filename: path::PathBuf,